+++ test/endpoint.js
@@ -68,6 +68,7 @@ describe('metrics', function() {
     require('./metrics/distance/overlap.js');
     require('./metrics/distance/prefix.js');
     require('./metrics/distance/smith-waterman.js');
+    require('./metrics/distance/suffix.js');
   });
 });