+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Adding the `distance/metrics/length` namespace.
 * Adding the `distance/metrics/prefix` namespace.
 * Adding the `distance/metrics/smith-waterman` namespace.
+* Adding the `distance/metrics/suffix` namespace.
 * Adding the `phonetics/french/fonem` namespace.
 * Adding the `regexp` namespace.
 * Adding the `stemmers/french/carry` namespace.